Add BirthDate property with validation to User class
Introduced a new `BirthDate` property to the `User` class within the `Library` namespace. The property includes a getter and a setter. The setter contains a validation check to ensure that the birth date cannot be set to a future date. If the provided birth date is in the future, an `ArgumentException` is thrown with the message "Birth-Date cannot be in the future!".