Add Email and Password properties to User class
Introduced two new properties, `Email` and `Password`, to the
`User` class within the `Library` namespace in the `User.cs`
file. These properties include getter and setter methods,
allowing for the retrieval and assignment of the `_email` and
`_password` fields, respectively. This enhancement enables
other parts of the code to access and modify the user's email
and password.