Add navigation from update page to main menu via link
Added an event handler for the `LinkClicked` event of `linkLabel1` in `Admin_Update_Page.Designer.cs`. Implemented the `linkLabel1_LinkClicked` method in `Admin_Update_Page.cs` to instantiate and show the `Admin_MainMenu` form, then close the current form. This allows users to navigate back to the main menu from the update page.